+++ meteoio/meteostats/libinterpol1D.cc
@@ -310,7 +310,7 @@ double Interpol1D::variance(const std::vector<double>& X)
 			sum3 = sum3 + (value - mean);
 		}
 	}
-	const double variance = (sum2 - sum3*sum3/count) / (count - 1);
+	const double variance = (sum2 - sum3*sum3/static_cast<double>(count)) / static_cast<double>(count - 1);
 	return variance;
 }
